[ethcore error]: introduce `BlockErrorWithData`
The main motivation behind this change is to include `raw rlp block bytes` in `BlockError`.
Instead of at runtime have to ensure that it is returned every time a `BadBlock` is detected.

This commit doesn't solve it completly because in a few major traits (Engine, ValidatorSet, EpochVerfier)
returns `EthcoreError`. Thus, that is why we have `Option<Bytes>` in `BlockErrorWithData` and still have to
ensure that it is not `None` at runtime which is unfortunate. Related to this, all functions/methods that I
have detected that could return more precise error have been changed.

Also, this commit includes a bunch of `nit fixes` that are unrelated to #11403.